fix: auto timezone: handle if timezone server is not reachable
In case the timezone server is not reachable we use know the
last known timezone instead of falling back to timezone +/-0.

When the we receive a answer from timezone server, it will be
checked if the timezone has changed. Only iff the timezone has
changed the new value will be stored in the config.